S3 ICC fix + SEO overhaul + code quality + wiki update
S3 fix: KCC-20 input ownership validation (ICC pattern) in daglock_krc20.sil
- validateKcc20Input() helper using readInputStateWithTemplate + OpCovInputIdx
- Gated by kcc20TemplatePrefixLen != 0 (dev/test mode bypass)
- New template hash: ae0946e4a9bd4a7585e6bf9135de38083cb11c85

Code quality: remove .unwrap() in production, fix fee denominator type
- All 9 .unwrap() calls replaced (UUID gen, mutex, treasury key)
- FEE_DENOMINATOR changed from u64 to i64 for type consistency
- Flaky crypto tests fixed (mutex poisoning recovery)

SEO overhaul: History API routing + meta tags + PWA
- Hash router → History API (real URLs for Google indexing)
- react-helmet-async with per-page unique titles + descriptions
- OG tags, Twitter Cards, JSON-LD structured data
- Sitemap + robots.txt
- PWA manifest + service worker via vite-plugin-pwa
- App icons (192, 512) + OG social card image (1200x630)

Infrastructure: VPS hardened, wRPC endpoints documented
- daglock user, LimitNOFILE=65536
- toggle-wrpc.sh script for switching modes
- Laptop kaspad setup script (patched for testnet-12)
- Working mainnet wRPC: troy.kaspa.stream (borsh)

Wiki: updated with S3 fix docs, new glossary terms (ICC, Borsh, PNN)

+++ wasm-sdk/src/lib.rs
@@ -179,7 +179,7 @@ pub fn verify_template_match(script_hex: &str, template_hash_hex: &str) -> Resul
 /// JSON string with: `{ amount_sompi, fee_sompi, fee_percentage }`
 #[wasm_bindgen]
 pub fn calculate_fee(amount_sompi: i64) -> Result<String, JsError> {
-    let fee_sompi = amount_sompi / daglock_shared::FEE_DENOMINATOR as i64;
+    let fee_sompi = amount_sompi / daglock_shared::FEE_DENOMINATOR;
 
     let result = serde_json::json!({
         "amount_sompi": amount_sompi,
